Add another slapping by Apple -Wall -Werror
(This message is unimportant to the final commit)

+++ Replicator/Pusher.cc
@@ -390,7 +390,7 @@ namespace litecore { namespace repl {
         if(response.isInteger()) {
             status = (int)response.asInt();
         } else if(auto dict = response.asDict(); dict) {
-            status = dict["status"].asInt();
+            status = (int)dict["status"].asInt();
             serverRevID = dict["rev"].asString();
         }
